Only check graph invariants after modification
Summary:
Given that invariants are only invalidated when the modification count
is changed, it suffices to only call `_maybeCheckInvariants` from within
`_markModification` (and the constructor). This eliminates a function
call and cache key check for many public accessor functions, but does
not present a substantial improvement: the best-of-three timing only
improved from 3.866s to 3.799s, which is not statistically significant.
The patch makes it less obvious that the invariants are always in fact
valid, so I recommend that it not be merged.

Test Plan:
Existing unit tests suffice.

wchargin-branch: graph-check-invariants-only-indirectly