Refactor `RelationalView`
As originally written, `RelationalView` methods return Entries, which
have pointers to other Entries' addresses, and the `RelationalView`
enables retrieving those corresponding entries by their addresses.

This is actually quite inconvenient in practice, as it forces the user
to deal with a lot of spurious null-checking. The retrieval methods
(rightly) report that they will possibly return null. However, in the
extremely common case where the user wants to look up e.g. the comments
on an issue, forcing them to explicitly null check is tedious and
unnecessary.

This change makes the test code more convenient, and will have a similar
impact on future callers. I changed the snapshot test somewhat, so that
it no longer snapshots the whole repository - that would now involve
snapshotting all of the data in the `RelationalView`.

Test plan:
`yarn travis`