Add a type field to `RelationalView` entries
This commit adds an explicit type field to the entries in
`RelationalView`. This is redundant (the exact same type is present on
`entry.address`) but we need these types for Flow to allow type
refinement on entries.

We also add the `Entry` type, the union of all the specific entries,
which is the type that we can now refine over thanks to the added types.

Test plan:
`yarn travis`