Update README (#482)
Rewrites the README to be a lot more concrete about what SourceCred is
doing, and to give more up-to-date information overall.

Test plan: The README includes a small code block with instructions for
turning on a local copy of the Cred Explorer. I ran this on my machine
(inside /tmp) and it worked locally.